Minor restructuring of services
