+++ scenarioo-client/app/shared/restServices.ts
@@ -1,0 +1,45 @@
+/* scenarioo-client
+ * Copyright (C) 2014, scenarioo.org Development Team
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import "./scenarioService.config";
+import './services/applicationStatus.service';
+import "./services/scenariooResource.service";
+import "./services/branchesResource.service";
+import "./services/buildImportStatesResource.service";
+import "./services/buildImportLogResource.service";
+import "./services/buildReimportResource.service";
+import "./services/buildImport.service";
+import "./services/pageVariant.service";
+import "./services/useCase.service";
+import "./services/fullTextSearch.service";
+import "./services/stepResource.service";
+import "./services/configResource.service";
+import "./services/useCasesResource.service";
+import "./services/scenarioResource.service";
+import "./services/objectsForTypeResource.service";
+import "./services/versionResource.service";
+import "./services/branchAliasResource.service";
+import "./services/customTabContentResource.service";
+import "./services/objectListResource.service";
+import "./services/objectIndexListResource.service";
+import "./services/useCasesResources.service";
+import "./services/labelConfigurationsListResource.service";
+import "./services/labelConfigurationsResource.service";
+import "./services/comparisonsResource.service";
+import "./services/comparisonLogResource.service";
+import "./services/comparisonCreateResource.service";
+import "./services/comparisonRecalculateResource.service";
