+++ scenarioo-client/app/shared/scenarioService.config.ts
@@ -1,0 +1,5 @@
+angular.module('scenarioo.services')
+    .config(function ($httpProvider) {
+        $httpProvider.defaults.headers.common.Accept = 'application/json';
+        $httpProvider.defaults.stripTrailingSlashes = false;
+    });
