+++ scenarioo-client/app/shared/services/applicationStatus.service.ts
