+++ scenarioo-client/app/shared/services/branchAliasResource.service.ts
@@ -1,0 +1,4 @@
+angular.module('scenarioo.services')
+    .factory('BranchAliasesResource', function (ScenariooResource) {
+        return ScenariooResource('/branchaliases', {}, {});
+    });
