+++ scenarioo-client/app/shared/services/buildImportLogResource.service.ts
@@ -1,0 +1,31 @@
+/* scenarioo-client
+ * Copyright (C) 2014, scenarioo.org Development Team
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+angular.module('scenarioo.services')
+
+    .factory('BuildImportLogResource', function ($http) {
+        return {
+            get: function (branchName, buildName, onSuccess, onError) {
+                var callURL = 'rest/builds/importLogs/' + encodeURIComponent(branchName) + '/' + encodeURIComponent(buildName);
+                $http({
+                    method: 'GET', url: callURL, headers: {
+                        'Accept': 'text/plain'
+                    }
+                }).success(onSuccess).error(onError);
+            }
+        };
+    });
