+++ scenarioo-client/app/shared/services/comparisonsResource.service.ts
@@ -1,0 +1,4 @@
+angular.module('scenarioo.services')
+    .factory('ComparisonsResource', function (ScenariooResource) {
+        return ScenariooResource('/comparisons', {}, {});
+    });
