+++ scenarioo-client/app/shared/services/configResource.service.ts
@@ -1,0 +1,5 @@
+angular.module('scenarioo.services')
+    .factory('ConfigResource', function (ScenariooResource) {
+        return ScenariooResource('/configuration', {});
+    });
+
