+++ scenarioo-client/app/shared/services/customTabContentResource.service.ts
@@ -1,0 +1,10 @@
+
+angular.module('scenarioo.services')
+    .factory('CustomTabContentResource', function (ScenariooResource) {
+        return ScenariooResource('/branches/:branchName/builds/:buildName/customTabObjects/:tabId',
+            {
+                branchName: '@branchName',
+                buildName: '@buildName',
+                tabId: '@tabId'
+            }, {});
+    });
