+++ scenarioo-client/app/shared/services/labelConfigurationsListResource.service.ts
@@ -1,0 +1,4 @@
+angular.module('scenarioo.services')
+    .factory('LabelConfigurationsListResource', function (ScenariooResource) {
+        return ScenariooResource('/labelconfigurations/list', {}, {});
+    });
