+++ scenarioo-client/app/shared/services/labelConfigurationsResource.service.ts
@@ -1,0 +1,4 @@
+angular.module('scenarioo.services')
+    .factory('LabelConfigurationsResource', function (ScenariooResource) {
+        return ScenariooResource('/labelconfigurations', {}, {'query': {isArray: false}});
+    });
