+++ scenarioo-client/app/shared/services/objectsForTypeResource.service.ts
@@ -1,0 +1,8 @@
+angular.module('scenarioo.services')
+    .factory('ObjectsForTypeResource', function (ScenariooResource) {
+        return ScenariooResource('/branches/:branchName/builds/:buildName/objects/service',
+            {
+                branchName: '@branchName',
+                buildName: '@buildName'
+            }, {});
+    });
