+++ scenarioo-client/app/shared/services/pageVariant.service.ts
@@ -1,0 +1,14 @@
+angular.module('scenarioo.services')
+
+    .factory('PageVariantService', function (ScenariooResource, $q) {
+        var pageVariantService = ScenariooResource('/branch/:branchName/build/:buildName/search/pagevariants/',
+            {
+                branchName: '@branchName',
+                buildName: '@buildName'
+            }, {});
+
+        pageVariantService.getPageVariantCount = getPromise($q, function (parameters, fnSuccess, fnError) {
+            return pageVariantService.get(parameters, fnSuccess, fnError);
+        });
+        return pageVariantService;
+    })
