+++ scenarioo-client/app/shared/services/scenariooResource.service.ts
@@ -1,0 +1,26 @@
+/* scenarioo-client
+ * Copyright (C) 2014, scenarioo.org Development Team
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+angular.module('scenarioo.services')
+    /**
+     * All resources in Scenarioo must be based on this ScenariooResource.
+     */
+    .factory('ScenariooResource', function ($resource) {
+        return function (url, paramDefaults, actions) {
+            return $resource('rest' + url, paramDefaults, actions);
+        };
+    })
