+++ scenarioo-client/app/shared/services/versionResource.service.ts
@@ -1,0 +1,4 @@
+angular.module('scenarioo.services')
+    .factory('VersionResource', function (ScenariooResource) {
+        return ScenariooResource('/version', {}, {});
+    });
